Try fixing flaky functional tests
