+++ setup.cfg
@@ -23,8 +23,6 @@ line_length = 88
 
 [tool:pytest]
 junit_family = legacy
-markers =
-    gui_serial: GUI tests that should run serially on macOS to avoid xdist worker crashes
 addopts =
     --cov-report=
     --cov-report=html
